+++ Source/image_header.hpp
@@ -18,8 +18,10 @@ namespace tga
 		uint8_t colorMapBitDepth;
 		uint16_t originX;
 		uint16_t originY;
-		uint16_t width;
-		uint16_t height;
+		//uint16_t width;
+		//uint16_t height;
+		int width;
+		int height;
 		uint8_t pixelBitDepth;
 		uint8_t imageDescriptor;
 		std::string imageId;
